Iss34 Resolved

+++ peonordersystem/ConfirmationSystem.py
@@ -119,7 +119,8 @@ def order_confirmed(order_name, priority_list,
     is included for purposes of printing and retrieving
     accurate data.
     """
-    # TODO send priority_list, order_list to kitchen
+    print_order(order_name, non_priority_list,
+                priority_list=priority_list)
 
     curr_directory = directory + 'confirmed/'
     order_name = standardize_confirm_file_name(order_name)
@@ -141,9 +142,11 @@ def remove_order_confirmed_file(order_name):
     """
     curr_directory = directory + 'confirmed/'
     order_name = standardize_confirm_file_name(order_name)
-    
-    os.remove(curr_directory + order_name)
 
+    f_name = curr_directory + order_name
+
+    if os.path.isfile(f_name):
+        os.remove(curr_directory + order_name)
 
 def checkout_confirmed(order_name, orders, order_list):
     """Generates the necessary checkout files
@@ -173,9 +176,58 @@ def checkout_confirmed(order_name, orders, order_list):
     curr_file.write(obj_info)
 
     for order in orders:
-        # TODO send order to checkout printer
-        print order
-        pass
+        print_check(order_name, order)
+
+
+def print_order(order_name, order_list, priority_list=None):
+    """Send the given order to the order
+    printer. If given a priority order the
+    priority order is given special priority
+    status and printed first. The remaining
+    order is printed last.
+
+    @param order_name: str representation of
+    the order.
+
+    @param order_list: list of MenuItem objects
+    that represents the order to be sent to the
+    order printer.
+
+    @keyword priority_list: list of MenuItem
+    objects that represents the priority order to
+    be sent to the order printer.
+
+    @return: None
+    """
+    #TODO print order to order printer
+    print 'TO ORDER'
+    print order_name
+    if priority_list:
+        print priority_list
+
+    print order_list
+    print ''
+    pass
+
+
+def print_check(order_name, order_list):
+    """Send the given order to the check
+    printer.
+
+    @param order_name: str representation
+    of the orders name
+
+    @param order_list: list of MenuItem objects
+    that is to be printed to the check printer.
+
+    @return: None
+    """
+    #TODO send order to checkout printer
+    print 'TO CHECK'
+    print order_name
+    print order_list
+    print''
+    pass
 
 
 def unpack_information(curr_directory):
